+++ lib-multisrc/fuzzydoodle/src/eu/kanade/tachiyomi/multisrc/fuzzydoodle/Filters.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.multisrc.cloudrecess
+package eu.kanade.tachiyomi.multisrc.fuzzydoodle
 
 import eu.kanade.tachiyomi.source.model.Filter
 import okhttp3.HttpUrl
