+++ src/ar/hentaislayer/build.gradle
@@ -1,7 +1,7 @@
 ext {
     extName = 'Hentai Slayer'
     extClass = '.HentaiSlayer'
-    themePkg = 'cloudrecess'
+    themePkg = 'fuzzydoodle'
     overrideVersionCode = 2
     isNsfw = true
 }
