+++ src/en/fleksyscans/src/eu/kanade/tachiyomi/extension/en/fleksyscans/FleksyScans.kt
@@ -1,5 +1,5 @@
 package eu.kanade.tachiyomi.extension.en.fleksyscans
 
-import eu.kanade.tachiyomi.multisrc.cloudrecess.CloudRecess
+import eu.kanade.tachiyomi.multisrc.fuzzydoodle.FuzzyDoodle
 
-class FleksyScans : CloudRecess("FleksyScans", "https://flexscans.com", "en")
+class FleksyScans : FuzzyDoodle("FleksyScans", "https://flexscans.com", "en")
